Fixed the import route and some renaming folders
