+++ server/src/controllers/registration/completeProfile.js
