+++ server/src/controllers/registration/verifyEmail.js
